requirements/local.txt
-r ./base.txt

Werkzeug==0.15.2  # https://github.com/pallets/werkzeug
ipdb==0.12  # https://github.com/gotcha/ipdb
Sphinx==2.0.1  # https://github.com/sphinx-doc/sphinx
psycopg2-binary==2.8  # https://github.com/psycopg/psycopg2

# Testing
# ------------------------------------------------------------------------------
mypy==0.701  # https://github.com/python/mypy
pytest==4.4.1  # https://github.com/pytest-dev/pytest
pytest-sugar==0.9.2  # https://github.com/Frozenball/pytest-sugar

# Code quality
# ------------------------------------------------------------------------------
flake8==3.7.5  # https://github.com/PyCQA/flake8
coverage==4.5.3  # https://github.com/nedbat/coveragepy
black==19.3b0  # https://github.com/ambv/black
pylint-django==2.0.9  # https://github.com/PyCQA/pylint-django

# Django
# ------------------------------------------------------------------------------
factory-boy==2.11.1  # https://github.com/FactoryBoy/factory_boy

django-debug-toolbar==1.11  # https://github.com/jazzband/django-debug-toolbar
django-extensions==2.1.6  # https://github.com/django-extensions/django-extensions
django-coverage-plugin==1.6.0  # https://github.com/nedbat/django_coverage_plugin
pytest-django==3.4.8  # https://github.com/pytest-dev/pytest-django
django-multiselectfield==0.1.8
django-nose==1.4.6p
django-taggit==1.1.0
